requirements.txt
fastapi
uvicorn
torch
Pillow
requests
python-dotenv
supabase
openai-clip